setting up message based command executor
