+++ activiti-services/activiti-services-core/src/main/java/org/activiti/services/core/commands/CommandExecutor.java
@@ -1,0 +1,10 @@
+package org.activiti.services.core.commands;
+
+import org.activiti.services.core.model.commands.Command;
+
+public interface CommandExecutor {
+
+    Class getHandledType();
+
+    void execute(Command cmd);
+}
