+++ activiti-services/activiti-services-core/src/main/java/org/activiti/services/core/pageable/sort/ProcessDefinitionSortApplier.java
@@ -13,7 +13,7 @@
  *
  */
 
-package org.activiti.services.core.tests.pageable.sort;
+package org.activiti.services.core.pageable.sort;
 
 import java.util.HashMap;
 import java.util.Map;
