+++ activiti-services/activiti-services-core/src/test/java/org/activiti/services/core/pageable/ProcessDefinitionSortApplierTest.java
@@ -1,25 +1,24 @@
 /*
- * Copyright 2017 Alfresco, Inc. and/or its affiliates.
- *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
  * You may obtain a copy of the License at
  *
- *       http://www.apache.org/licenses/LICENSE-2.0
+ *      http://www.apache.org/licenses/LICENSE-2.0
  *
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ *
  */
 
-package org.activiti.services.core.tests.pageable;
+package org.activiti.services.core.pageable;
 
 import org.activiti.engine.ActivitiIllegalArgumentException;
-import org.activiti.engine.impl.TaskQueryProperty;
-import org.activiti.engine.task.TaskQuery;
-import org.activiti.services.core.tests.pageable.sort.TaskSortApplier;
+import org.activiti.engine.impl.ProcessDefinitionQueryProperty;
+import org.activiti.engine.repository.ProcessDefinitionQuery;
+import org.activiti.services.core.pageable.sort.ProcessDefinitionSortApplier;
 import org.junit.Before;
 import org.junit.Rule;
 import org.junit.Test;
@@ -33,10 +32,10 @@ import static org.activiti.services.core.tests.utils.MockUtils.selfReturningMock
 import static org.mockito.Mockito.*;
 import static org.mockito.MockitoAnnotations.initMocks;
 
-public class TaskSortApplierTest {
+public class ProcessDefinitionSortApplierTest {
 
     @InjectMocks
-    private TaskSortApplier sortApplier;
+    private ProcessDefinitionSortApplier sortApplier;
 
     @Rule
     public ExpectedException expectedException = ExpectedException.none();
@@ -47,9 +46,9 @@ public class TaskSortApplierTest {
     }
 
     @Test
-    public void applySort_should_oder_by_task_id_asc_by_default() throws Exception {
+    public void applySort_should_oder_by_process_definition_id_asc_by_default() throws Exception {
         //given
-        TaskQuery query = selfReturningMock(TaskQuery.class);
+        ProcessDefinitionQuery query = selfReturningMock(ProcessDefinitionQuery.class);
         PageRequest pageRequest = PageRequest.of(0,
                                                  10);
 
@@ -58,22 +57,22 @@ public class TaskSortApplierTest {
                               pageRequest);
 
         //then
-        verify(query).orderByTaskId();
+        verify(query).orderByProcessDefinitionId();
         verify(query).asc();
     }
 
     @Test
     public void applySort_should_use_the_criteria_defined_by_pageable_object() throws Exception {
         //given
-        TaskQuery query = selfReturningMock(TaskQuery.class);
-        Sort.Order processDefinitionOrder = new Sort.Order(Sort.Direction.ASC,
-                                                           "name");
-        Sort.Order processInstanceOrder = new Sort.Order(Sort.Direction.DESC,
-                                                         "id");
+        ProcessDefinitionQuery query = selfReturningMock(ProcessDefinitionQuery.class);
+        Sort.Order deploymentOrder = new Sort.Order(Sort.Direction.ASC,
+                                                    "deploymentId");
+        Sort.Order processDefinitionOrder = new Sort.Order(Sort.Direction.DESC,
+                                                           "id");
         PageRequest pageRequest = PageRequest.of(0,
                                                  10,
-                                                 Sort.by(processDefinitionOrder,
-                                                         processInstanceOrder));
+                                                 Sort.by(deploymentOrder,
+                                                         processDefinitionOrder));
 
         //when
         sortApplier.applySort(query,
@@ -81,16 +80,16 @@ public class TaskSortApplierTest {
 
         //then
         InOrder inOrder = inOrder(query);
-        inOrder.verify(query).orderBy(TaskQueryProperty.NAME);
+        inOrder.verify(query).orderBy(ProcessDefinitionQueryProperty.DEPLOYMENT_ID);
         inOrder.verify(query).asc();
-        inOrder.verify(query).orderBy(TaskQueryProperty.TASK_ID);
+        inOrder.verify(query).orderBy(ProcessDefinitionQueryProperty.PROCESS_DEFINITION_ID);
         inOrder.verify(query).desc();
     }
 
     @Test
     public void applySort_should_throw_exception_when_using_invalid_property_to_sort() throws Exception {
         //given
-        TaskQuery query = selfReturningMock(TaskQuery.class);
+        ProcessDefinitionQuery query = selfReturningMock(ProcessDefinitionQuery.class);
         Sort.Order invalidProperty = new Sort.Order(Sort.Direction.ASC,
                                                     "invalidProperty");
         PageRequest pageRequest = PageRequest.of(0,
