+++ activiti-services/activiti-services-rest/src/main/java/org/activiti/services/rest/controllers/TaskController.java
@@ -21,8 +21,8 @@ import org.activiti.engine.TaskService;
 import org.activiti.services.core.model.Task;
 import org.activiti.services.core.model.commands.CompleteTaskCmd;
 import org.activiti.services.core.model.converter.TaskConverter;
-import org.activiti.services.core.tests.pageable.AuthenticationWrapper;
-import org.activiti.services.core.tests.pageable.PageableTaskService;
+import org.activiti.services.core.pageable.AuthenticationWrapper;
+import org.activiti.services.core.pageable.PageableTaskService;
 import org.activiti.services.rest.resources.TaskResource;
 import org.activiti.services.rest.resources.assembler.TaskResourceAssembler;
 import org.springframework.beans.factory.annotation.Autowired;
