+++ activiti-spring-boot/spring-boot-starters/activiti-spring-boot-starter/src/test/java/org/activiti/runtime/TasksIT.java
@@ -28,8 +28,8 @@ import org.activiti.services.core.model.ProcessDefinition;
 import org.activiti.services.core.model.ProcessInstance;
 import org.activiti.services.core.model.Task;
 import org.activiti.services.core.model.commands.CompleteTaskCmd;
-import org.activiti.services.core.tests.pageable.AuthenticationWrapper;
-import org.activiti.services.core.tests.pageable.PageableTaskService;
+import org.activiti.services.core.pageable.AuthenticationWrapper;
+import org.activiti.services.core.pageable.PageableTaskService;
 import org.activiti.services.rest.controllers.TaskController;
 import org.junit.Before;
 import org.junit.Test;
